This function executes all tests for drift between two datasets / models

Currently three checks are implemented, covariate drift, residual drift and model drift.

check_drift(model_old, model_new, data_old, data_new, y_old, y_new,
  predict_function = predict, max_obs = 100, bins = 20,
  scale = sd(y_new, na.rm = TRUE))

Arguments

model_old
model created on historical / `old`data

model_new
model created on current / `new`data

data_old
data frame with historical / `old` data

data_new
data frame with current / `new` data

y_old
true values of target variable for historical / `old` data

y_new
true values of target variable for current / `new` data

predict_function
function that takes two arguments: model and new data and returns numeric vector with predictions, by default it's `predict`

max_obs
if negative, them all observations are used for calculation of PDP, is positive, then only `max_obs` are used for calculation of PDP

bins
continuous variables are discretized to `bins` intervals of equal sizes

scale
scale parameter for calculation of scaled drift

Value

This function is executed for its side effects, all checks are being printed on the screen. Additionaly it returns list with particualr checks.

Examples

 library("DALEX")

 library("ranger")
 predict_function <- function(m,x,...) predict(m, x, ...)$predictions
 model_old <- ranger(m2.price ~ ., data = apartments)
 model_new <- ranger(m2.price ~ ., data = apartments_test)
 check_drift(model_old, model_new,
                  apartments, apartments_test,
                  apartments$m2.price, apartments_test$m2.price,
                  predict_function = predict_function)

#>    -------------------------------------
#>                   Variable  Shift
#>   -------------------------------------
#>                   m2.price    4.9  
#>          construction.year    6.0  
#>                    surface    6.8  
#>                      floor    4.9  
#>                   no.rooms    2.8  
#>                   district    2.8  
#>    -------------------------------------
#>                   Variable  Shift
#>   -------------------------------------
#>                  Residuals   34.7  **
#>    -----------------------------------------------
#>                   Variable    Shift  Scaled
#>   -----------------------------------------------
#>                      floor    75.69     8.4  
#>                   no.rooms   107.66    12.0  .
#>                    surface    86.66     9.6  
#>                   m2.price   126.46    14.0  .
#>          construction.year   132.76    14.7  .